+++ mycli/packages/special/iocommands.py
@@ -17,6 +17,7 @@ TIMING_ENABLED = False
 use_expanded_output = False
 PAGER_ENABLED = True
 tee_file = None
+once_file = None
 
 @export
 def set_timing_enabled(val):
@@ -251,10 +252,8 @@ def execute_system_command(arg, **_):
     except OSError as e:
         return [(None, None, None, 'OSError: %s' % e.strerror)]
 
-@special_command('tee', 'tee [-o] filename',
-                 'write to an output file (optionally overwrite using -o)')
-def set_tee(arg, **_):
-    global tee_file
+
+def parseargfile(arg):
     if arg.startswith('-o '):
         mode = "w"
         filename = arg[3:]
@@ -265,8 +264,16 @@ def set_tee(arg, **_):
     if not filename:
         raise TypeError('You must provide a filename.')
 
+    return {'file': filename, 'mode': mode}
+
+
+@special_command('tee', 'tee [-o] filename',
+                 'write to an output file (optionally overwrite using -o)')
+def set_tee(arg, **_):
+    global tee_file
+
     try:
-        tee_file = open(filename, mode)
+        tee_file = open(**parseargfile(arg))
     except (IOError, OSError) as e:
         raise OSError("Cannot write to file '{}': {}".format(e.filename, e.strerror))
 
@@ -291,3 +298,30 @@ def write_tee(output):
         tee_file.write(output)
         tee_file.write(u"\n")
         tee_file.flush()
+
+
+@special_command('\\once', '\\o [-o] filename', 'Output for the next SQL command only to FILENAME', aliases=('\\o', ))
+def set_once(arg, **_):
+    global once_file
+
+    once_file = parseargfile(arg)
+
+    return [(None, None, None, "")]
+
+
+@export
+def write_once(output):
+    global once_file
+    if output and once_file:
+        try:
+            f = open(**once_file)
+        except (IOError, OSError) as e:
+            once_file = None
+            raise OSError("Cannot write to file '{}': {}".format(
+                e.filename, e.strerror))
+
+        with f:
+            f.write(output)
+            f.write(u"\n")
+
+        once_file = None